refactor: isolate profile export security boundary
